Mark known sha1 mismatches differently
There are a few remaining sha1 mismatches in Mozilla repositories as well
as the Mercurial repository. While they are not handled yet, they are on
old enough changesets that they would not create any problem. While it is
not clear how they will be fixed, it will likely not involve changing
the git commit sha1s anyways, so it is safe to ignore them for now.