Do not require GitHgStore.manifest callers to handle incremental manifests themselves
For repositories with a lot of files, generating a manifest is a slow
operation.
In most cases, there are way less changes between changesets than there
are files in the repository, so it is much faster to generate a manifest
from a previously generated manifest, by applying the differences between
the corresponding trees.
This was required to be handled by callers, but it works equally if it's
taken care of by the callee, and ends up being simpler.