Generate changeset patches generically
For differences between a mercurial changeset generated from a git commit +
the metadata stored in a note and the actual changeset, we currently have
a generic "patch" field in the note that allows arbitrary differences.
But at the moment, the patch is only generated for the known difference
in timezone offset granularity. In practice there aren't other differences
that matter, but for repo grafting, there are going to be more differences,
so generate the patch generically.