Add minimalistic and optimistic compatibility with hg-git on pushes with graft
The gecko-dev repository is filled with hg-git, which has many differences
with cinnabar. When trying to follow the gecko-dev repo with grafts, and
pushing to mozilla mercurial repositories, the corresponding commits in
gecko-dev won't match the local ones because of these differences.
But in cases not involving hg-git adding a '--HG--' section (mostly, no file
copy/rename), the difference is only a carriage return at the end of the
commit message.
By stripping the final carriage return when pushing, we can make hg-git
create the same commit as the one that was pushed, as long as it doesn't
add a '--HG--' section.
However, that is not a viable option in the general case, because it breaks
git->hg->git round trips with cinnabar. It is however fine-ish to do when
grafting and one parent commit is detected to have that carriage return
discrepancy.