Fix regression from e4db461
e4db461 changed how the commit command for fast-import was invoked such that
a "from" was always given, but that happens to trigger a suboptimal behavior
of fast-import, where it will, presumably, always trigger an actual read
of the corresponding tree, even if the from matches the current value of the
ref.

This, in turn, is likely to trigger a flush of the fast-import pack,
which is known to be extra slow on OSX. So while the regression is in the
order of 20% on Linux, it's more in the order of 400~500% on OSX.

So if we do know that the parent is the same as the ref value, we don't emit
a "from" anymore.