remove mercurial import from cinnabar/githg.py
The only thing githg.py needs the mercurial libraries for is generating
diffs of the format used in bundles.  For now we can just copy the two
functions in mercurial/pure/bdiff.py that are necessary to generate this
format of diff.  This may be somewhat slower than using the C
implementation of bdiff, however it should make it easier to switch to
generating bundles entirely in C.