Allow to store both split lines and full data in a GeneratedManifestInfo
To reduce the amount of memory used by a single GeneratedManifestInfo instance,
it only stored either the split lines or the full joined data, but never both
at the same time. Unfortunately, the way the class is used makes us alternating
between wanting one or the other, which skyrocketted the use of ManifestLine
instances when resplitting the full joined data into individual lines. This
also consumed a lot of time to re-split/re-join the data.

Allowing to store both split and joined data at the same time reduces both
memory usage and time spent on large pushes of large trees by a fair amount.