Don't use a --cat-blob-fd for fast-import
This used to be necessary in the old times where we'd send `progress`
commands to the fast-import process, so as to separate the output
from other commands such as `ls` or `cat-blob` from the output from
`progress` which was meant to users.

This simplifies the setup, and doesn't rely on file handle inheritance,
which is a nightmare to setup properly on Windows.