Allow to override the path to the helper through the environment or a config
Before, we could only disable the helper with GIT_CINNABAR_NO_HELPER.
This environment variable is now ignored, and instead, the full path to
the helper can be given through the GIT_CINNABAR_HELPER environment
variable or the cinnabar.helper config.

An empty value disables the helper, having the same effect as
GIT_CINNABAR_NO_HELPER had in the past.