Change which commit is being replaced for grafts
When doing a graft with non-grafted early history, both lines of history
are joined together to form the whole mercurial history with a
git-replace'ish ref. When first adding graft support, the replaced commit
was made to be the parent of the first grafted commit. With this change,
the replaced commit is the first grafted commit.

This means new grafted clones will, starting from this commit, be different,
but both views of the world are still valid as far as cinnabar is concerned
(all operations should still work with both kinds of clones, and fsck should
be happy with both).