Upgrade previous graft to the new scheme
(see parent commit for the new scheme)

As this requires detection of the current state of affairs, and there is
no directly detectable difference, we also need to do some metadata
change at the same time, which, come to think of it, is necessary
anyways: we now will also store the cinnabar/replace refs in the tree
under the cinnabar/metadata ref. This means all the metadata is now
also under the cinnabar/metadata and all the refs that are required for
the helper or some other git tool can be recreated from there. [We
however don't fully handle this yet because of the complications the
upgrade path already adds]