Rely on git's own handling of replace refs
Instead of trying to do things manually, which is error-prone, especially for
everything related to revwalking, rely on git handling it. Unfortunately,
support for replace refs "namespaces" is only available in git 2.6, so we
go through great lengths to ensure the whole thing still works with versions
earlier than that.

This also has the advantage of making cinnabar not confused by replace refs
that would be there for other reasons.