Make cinnabar.graft=only better, and make it the only option for grafting
There are a couple things that were wrong with cinnabar.graft=only:
- It would not fail if no changeset is grafted at all
- It would fail an entire fetch if one of the fetched heads is not
  grafted while (part of) its history is grafted.

This change makes grafting skip the part of history that is not
graftable and makes things work out somehow for git when fetching that
by only exposing the list of things that were successfully grafted.
There are a few caveats and UX flaws, but it's a better state than
what was available.