"Support" mercurial nodes with only a second parent
This is an example of a real life changeset that has this property:
changeset:   39564:b3dad02839a4
parent:      -1:000000000000
parent:      39557:b06226b66091

It comes from the pypy repository on bitucket. The corresponding manifest
inherits the weirdness.

Git cannot represent this kind of parenthood in its dag. Well, maybe it
does, but trying to create one through the standard means yields the
following message: "fatal: Non-fast-forward commit does not make sense
into an empty head". (I haven't actually tried harder to create one)

But the parent order doesn't actually matter to mercurial: with the
parents inverted, the changeset sha1 doesn't change (it's a property of
mercurial's data model).

Considering how reaching such changesets is convoluted (you need to
get your work tree on the null changeset and then merge something else
on it), and that the only effect of ignoring the order is that pushing
the changeset to some other repository not containing it already would
make that repository see a different order, just go with ignoring the
order.