Support the null manifest
When starting a new branch off the null changeset, you can end up with a
null manifest associated with it. We need to be able to accept such a
changeset, be able to push it properly, and if such a changeset is the
only one in a repository, we need to handle the lack of manifest dag in
our metadata.