Fix upgrading graft replace ref
d483c644 made fsck upgrade to the new graft scheme, but in some cases,
it was removing the replace refs, breaking mercurial metadata for early
changesets.

This was only a problem when upgrading from the initial graft support
before the release of 0.3.0 (that wasn't part of any release) to the
graft support from 0.3.0.