Use a different strategy to avoid munmap/mmap cycles
Fast-import has many ways in which it will trigger munmap/mmap cycles,
and those are expensive on OSX. The most problematic case since 0.3 is
that storing manifests whose parents are not the previous manifest will
trigger such a cycle, and many ways around it do too. In fact, the only
way to prevent the cycles is to still create commits linearly for all
the manifests, and then create a new set of commits for manifests,
matching what we want in the end. But on large repositories, this would
leave, after a clone, a large number of unreachable objects, which git
gc would happily store as loose objects, and that too is not desirable.
So we also keep those commits somewhere reachable from a ref under
refs/cinnabar.