Remove the "batch" processing of manifests
This was used to prevent mmap/munmap cycles, but it's too complex a system,
and requires even more complexity to keep up with generaldelta bundles.

We'll instead rely on different tricks, combined with upstream
improvements to fast-import (or, alternatively, through our own helper)

The immediate consequence is that cloning big repositories is going to
take a very long time on OSX until those things are in place.

Note this however keeps the batch reading of the git trees after all
manifests have been stored, and keeps it in GitHgStore because it's
better there than in getbundle.