Keep the order of remote heads in the BranchMap
It used to be necessary to reverse them because of how tags were
handled, but it's not necessary now that tags are handled the same was
as in mercurial, which doesn't involve heads order.

OTOH, the order of the remote heads matters when pushing, and a bad
order can lead to the remote complaining that the repository changed
while preparing changes.