Don't delay storing manifests and files when bundling
AFAIR, this used to be necessary because fast-import would error out
when doing the rollback, which is not very user-friendly.

Now that we use the helper instead of fast-import, and that it silently
drops things on rollback instead of erroring out, we can safely use the
helper fast-import commands during the bundle instead of at close time.