Don't keep new metadata in memory when bundling
Now that we can generate the metadata from scratch once it's created, we
can drop the in-memory representations so as to save plenty of memory.
This results in bundling doing more things than before, but this also
allows to do larger bundles with less memory.

For instance, on Mozilla repositories, due to the manifest size, it was
common for bundling during pushes to require several gigabytes of
memory. While now possibly slower, pushes of the same scale should
not require more than a few hundred megabytes of memory.