Queue all commands given to FastImportCommitHelper
Recent changes have removed some level of caching that was happening on
the python side, which results in some places trying to query the helper
while processing a fast-import commit, which the helper can't handle at
the moment.

So, avoid fast-import commit commands and helper commands getting
intermixed by accumulating all the commands emitted to a
FastImportCommitHelper and sending them to the helper all at once.