Make diff-tree faster on manifest trees
Manifest trees contain commit references, and under normal
circumstances, git would use those for submodules. It turns out that
the submodule handling part of git reacts to them by trying to read
a config, which makes it try to read the null sha1 for each
diff-tree operation we do, which triggers at least 2 system calls per
alternate per diff-tree, which adds up pretty quickly and can be
responsible for a significant overhead (especially in the upcoming
use of diff-tree to feed the repository upgrade faster than what is
done currently).

So we add --ignore-submodules=dirty to avoid this configuration loading.

Also, since git doesn't know if the given sha1:path given to diff-tree
are committishes or local paths, it checks if the files exist in the
work tree too, which also adds to the overhead.

So we add -- to indicate that everything appearing before is flags or
committishes, not paths.