[fsck] Change how files metadata upgrade is done
Before this change, we scanned the hg2git branch in search for blobs,
read all of them, extracted metadata from there based on whether they
started with '\1\n', and stored the metadata separately, while computing
the sha1 of the remainder, and crossing fingers that there were no
corruptions, and that it matched the actual git blob in the
corresponding git commit.

Instead of this kind of brute-force approach, we now scan manifest trees
for files, allowing us to know both the actual git blob in the
corresponding git commit and the raw mercurial data with the metadata
included. Which allows to validate that there is no corruption, and also
allows to skip any reading when there is no metadata and both match,
which is most of the time.

This makes the process of upgrading take 5 minutes instead of 13 on my
machine for a clone of multiple mozilla repositories. That's not
counting the subsequent fsck, but the new process is also more
deterministic to allow a decision to maybe skip the fsck in the future.

Anyways, this also makes the upgrade process prepared to take on
manifest upgrades at the same time.