Try to resolve branch heads from the store before the branchmap
b171b80 did some changes to how refs are gathered, and that broke grafts,
which effectively pull first to be able to fill the refs for the
remote-helper list command.

And since that pull happens, we might as well avoid git triggering an
actual pull by telling it the resolved sha1s straight away.