Make fsck adjust metadata when previous versions left some extra metadata around.
After 41416fb, the extra data checked in fsck for committer is not the
raw metadata, so it would lead to fsck reporting errors that are
actually false positives.

Before e305d5f, we were actually trying to fix those kind of metadata
discrepancies, but as they were supposed to be fixed by earlier
versions's fsck, they were replaced with errors. But as of 41416fb, the
condition triggering those errors is wrong.

Those checks were left "just in case", but it turns out there is
metadata that turns out to be unnecessary, and that fresh clones
wouldn't create. They are however unrelated to the errors that fsck
erroneously reports.

So, we remove these checks entirely, and instead, adjust the metadata to
only contain what a fresh clone would create. In practice, this removes
"author" information from a lot of places where the git author doesn't
match the hg author exactly, but where the Authorship conversion does.