Don't keep initialized chunks around
Now that the helper can get us raw mercurial objects while they're in
flight, we can avoid keeping them in memory. This will save some memory
on bundles using changegroup2, at the expense of CPU.

The reason for this change, however, is unrelated to either CPU or
memory.  It's necessary in order to make things work properly with the
refactored HgObjects, albeit in a simple way. We'll see if it matters to
save some CPU here later.