Separate out file changes tracking when bundling manifests
The list of files that end up in the changeset metadata is not
necessarily related to how the manifest was created. It was kind of an
abuse of the manifest class to store it there.

In order to refactor how manifests work, it's necessary to decouple
those.

Actually, I suspect the list of files attached to the changeset is wrong
when bundling a commit with no change (--allow-empty) or something like
that. So, there's a likely positive effect to this refactor anyways.