Set fast-import features from the helper itself
Now that the helper always wraps fast-import, we can just
unconditionally enable the features we want... which turns out to only
be "force". "ls" and "notes" actually do nothing (they just fail on
older fast-import, and we can't run that anymore since we always use
the helper)

"done" will be dealt with separately.