Automatically enable fast-import feature "done"
Now that we have full control over fast-import, we can enable the "done"
feature at any appropriate time, which we now do automatically when the
helper receives a blob or commit command. It would make sense to do that
also for the reset command, but this didn't happen from the python code,
there may be assumptions that they don't persist.