Use the helper to keep track of all hg2git mappings
This moves the memory usage of those mappings to the helper process
(which should be less than the equivalent memory usage from the python
structures), and as a side effect, avoids the crazy spike in memory use
when closing the GitHgStore (see https://glandium.org/blog/?p=3771)

This does, however, have its own spike in memory use, but on the helper
side, due to, as of now, the use of an intermediate fast-import tree
while iterating the notes tree. Eventually, we'll just create tree
objects directly.