[helper] Increase atom table size
The default atom table size in fast-import is too small for a repository
the size of mozilla-central. The new size is a trade-off between memory
size for small repos, and overhead for large ones.

After 49071ce, storing the hg2git mappings can create a lot of new nodes
because of the extra fanout. That floods to_atom(), which then dominates
profiles during the finalization phase. Eventually, this will be less of
a concern when we stop using tree_content_set and create the notes tree
directly, at which point to_atom() won't be called at all.