Avoid hg2git requests in GitHgStore.store_*
When hg2git information was on the python side, checking the information
wasn't forcing synchronicity between git-remote-hg and
git-cinnabar-helper. The only reason the checks are there is to avoid
storing again information that we already have. This also avoids
overwriting information in case the imported bundle is forged to create
such overwrite.

We can avoid the overwrite on the helper side, at the expense of still
storing the data, which, after the import is finished, will just be
dangling data and will be eliminated by git gc.

In case we were actually given something we already had for some reason,
fast-import will find the corresponding data existing in git, and will
not create new objects anyways, and the hg2git info will already be
pointing there, so there's no loss of information.