Fix manifests tracking from 6f6330d not to drop heads on incremental updates
While the heads are not referenced in metadata^2, they're still in the
metadata history and can't be pruned by git gc. Since those references
are not strictly necessary for the good functioning of cinnabar (as long
as the hg2git data is full, their lack won't cause problems), we won't
care to make people run fsck. We'll wait until the next necessary
metadata upgrade, which is planned before 0.5 anyways.