Avoid git2hg and cat-file requests in GitHgStore.add_head
To do so, we keep a cache of branches for in-flight changesets.
Ideally, we'd just get rid of tracking branches (we don't /really/ need
it), and move the heads tracking to the helper, but we actually do need
the branch tracking for grafts, until we come up with something better
to handle the situation described in remote_helper.py:210.