Increase the default pack depth to 50
In git-core's 618e613a70, 10 years ago, the default for pack depth used
for git-pack-objects and git-repack was changed from 10 to 50, while
leaving fast-import's default to 10.

There doesn't seem to be a reason besides oversight for the change not
having happened in fast-import as well.

Interestingly, fast-import uses pack.depth when it's set, and the
git-config manual says the default for pack.depth is 50. While the
git-fast-import manual does say the default depth is 10, the
inconsistency is also confusing.

For git-cinnabar, on large repositories, this can make a significant
difference in pack sizes.