Send File RevChunks to the helper
So far, handling RevChunks has been done on the python end. They were
applied to the delta node and then the resulting data was sent to the
helper for storage.

Instead of doing that, we now just send the RevChunks to the helper,
which proceeds with applying them to the delta node and store the
data. This is limited to files, though. Manifests and changesets will
require more work.

Interestingly, this ends up being much faster than previously,
especially for changegroupv2, despite the helper not being
multi-threaded. The code is also not optimized at all. Some things
suffer from e.g. duplication of effort due to limitations to the
fast-import "API" (which simply come from it not being designed as an
API).