Add a LRU cache for git2hg, hg2git and cat-file("commit")
The GitHgStore used to keep an indefinitely growing cache at least
git2hg and hg2git data, and that consumed a lot of memory. While
requesting the data from the helper is fast, the repeated requests for
the same changesets during fsck are making the whole operation longer
than it should. So we cache the most recently requested results.

We also add cat-file("commit") because fsck does that a lot too with the
same commits.