Store file metadata as blobs instead of gitlinks
This effectively reverts 9ab64e2, but only for file-meta, not for hg2git
data. The intent of 9ab64e2 was to make the hg2git data for files not
directly linked to the file data, which is already available through
the converted changesets, allowing the cinnabar metadata being more
easily shared. But at the same time, it made the file-meta, which is
available nowhere else, use gitlinks too, allowing git-gc to purge
them, leading to corruption further down the road.