[experiment] Send Manifest RevChunks to the helper
This is kept behind a cinnabar.experiments flag for now.

Similarly to how we send File RevChunks to the helper, we now can send
Manifest RevChunks, which are then treated much faster. Despite the code
not being optimized much, this already makes a significant difference
with the python equivalent (almost twice as fast).