Hardcode how notes are stored.
We've been using oid_object_info on what's enumerated by for_each_note
mostly to handle trees properly, but it is, in fact, just simpler to
hardcode that if the path ends with a /, we've been given a tree, and
if the note tree is hg2git, all objects are gitlinks, vs. blob in other
note trees.

A side effect is that this makes updating the hg2git tree not fail when
there are leftovers from grafting (metadata is there, but the actual
commit is not because it was dropped).

This might even be faster (not using oid_object_info).