Intercept exceptions and provide a human readable message for the users.
Exceptions are intercepted in the `cinnabar.util.run` method.
If an exception is raised with a message, it is printed out.

A "traceback" check is introduced to provide an option to see the
traceback.

A new module `cinnabar/exceptions.py` is added to define all specific
exceptions.

`Abort` is defined as a base exception for normal command aborts.

`HelperClosed` is now a RuntimeError.