Change the manifest format to avoid rejection by some git servers
One of the goals of the current format was to allow sharing the metadata
tree to possibly make clones faster by pulling from git rather than
mercurial. However, this fell short with the fixes for CVE-2018-11235
being a little too restrictive, and server-side fsck (enabled on github)
thus rejecting the .gitmodules files from the mozilla-central manifests.

So, to avoid the problem altogether, we prefix all paths with an
underscore in the manifest metadata.