Change how buffers are pre-allocated when generating manifests
When generating manifests, we were pre-allocating a buffer based on the
length of the previous manifest. But this makes essentially all
allocations have a different size, and this can blow up depending on
allocation patterns with glibc, as per https://glandium.org/blog/?p=3698
and https://sourceware.org/bugzilla/show_bug.cgi?id=23416.

A better approach is to base it on the allocation size that was used
for the previous buffer, which allows sizes to be normalized to a
smaller set of different values, effectively working around the glibc
fragmentation problem.