[helper] Properly handle empty files when creating git trees
Because an empty file can have the same sha1 as an empty manifest,
we never store empty files in the hg2git notes tree. But if there
*is* an empty manifest, doing an hg2git notes lookup will return
the manifest metadata sha1, which is not valid to use when creating
git trees during changeset import.

This was properly handled by the corresponding python code, and is
effectively a regression from 0.5.0b3.