Always pass a `from` to `commit` commands
Before 0df3245721 in git-core, using `from` could be expensive, and the
FastImport code was adapted to avoid them when possible. This made sense
to do when we were still using git fast-import, but now that everything
related to fast-import is dealt with by the helper, and that the helper
has long contained 0df3245721, it's not necessary anymore.