Use different types and functions for hg sha1s as much as possible.
Because git upstream is moving to support different sizes of object ids, and on
the long run, the hg and git sizes might differ (e.g. cloning sha1 hg with
sha256 git), we cannot continue to use the same types, considering many
functions attached to object ids rely on the object id size of the git
repository.