[store-changegroup] Read the whole bundle first
This is a first step in the direction of making the download as fast as
possible while keeping the import in parallel. This is obviously not
what happens here, as we're explicitly delaying the import after the
entire download, making the entire affair slower, but this lays the
ground for the parallelization work.

Also, hide progress from the python code. The C code will have to report
its own progress, but we'll implement that later.