Make the bufferized writer use a background thread.
We want downloads to happen as fast as possible and not wait on cinnabar
processing, which is known to incur long pauses on the download, which
can have bad consequences on network connections when the server silently
drops connections after a few minutes (yes, that happens).

So we move writing back the buffered data to a separate thread where
writing can block without blocking the download.

This can obviously end up using large amounts of memory, if the
downloaded bundle is large and the download is faster than the import.
So a possible next step would be to store the data to disk if it reaches
a certain amount.