Pass an argument to the helper
This allows to determine whether the helper is called for the wire
protocol or for data import. Which can be useful for wrappers (I've
wanted to e.g. rr record only the wire part), and will be useful for
the helper itself to expose only the capabilities appropriate for the
mode.