[fsck] Add a faster mode to check files
Currently, `fsck --files` is the only way to check files, and because it
checks every single one of them, it takes a long time. This adds a new
option (which will probably become the default), that only checks files
for a changeset close to current HEAD, the assumption being that if a
corruption that matters happened in the ancestry, it will be visible on
the tip. Future improvements may include checking all mercurial heads.