[fsck] Change how quick fsck works
So far, it was checking the closest changeset. Now it checks all heads,
but only checks those, not the entire dag.

Another notable difference, for now, is that it doesn't keep trying to
find broken things further after it found one. We'll restore this
feature after some optimizations.

Interestingly, on a "small" repository, like the mercurial mercurial
repository, this is faster than the previous --quick. On a larger
repository, like mozilla-unified, this is slower, because of the larger
number of heads. OTOH, it's "only" 1.5 times slower than plain
`git cinnabar fsck` on that repo, while it can find errors that only
`git cinnabar fsck --manifest --files` would catch, while being at
least an order of magnitude slower, which puts this new version of
--quick as an attractive future default.