Allow to import cinnabarclone bundles from grafted repos
When generating e.g. a cinnabarclone bundle for a grafted gecko-dev,
one would rather exclude everything from gecko-dev so that the bundle
is smaller, since gecko-dev was presumably already cloned. In that case,
the bundle is thin, and index-pack needs the --fix-thin command line
option.
Replace refs also need to be handled properly.