cinnabarclone: add byte-string prefixes to strings that interface with Mercurial
Mercurial assumes all strings in its codebase are byte-strings.
Extensions need to ensure any string that interfaces with Mercurial
internals are specified as byte strings to be compatible with
Python 3. So this commit adds byte-string prefixes to all the
relevant strings in `cinnabarclone`.

After making this change, integration tests in version-control-tools
no longer fail when `cinnabarclone` is loaded into Mercurial,
and it should be compatible with the beta Python 3 release of
Mercurial (5.0).