Don't leak note leaf nodes when setting notes to the same value
In some cases, different mercurial files have the same node-id and thus
the same git blob (e.g. different files with the same history, or more
simply, different files with the same initial content). We don't check
whether a note already exists because in most cases it would be a waste
of time. So we do end up calling add_note with the same pair of
key/value, which is then leaked.

This fixes it.