[CI] Build Windows helper against the curl static library
Somehow, in some cases we can end up with Access Violation (0xc0000005)
when using the libcurl from msys2 (despite having linked against that
one) but not when using the one from git. It should be more reliable to
statically link.

Do a dummy change to the helper code so that it gets a new hash and is
rebuilt.