Refactor hg_connect_http to use readers for HTTP responses
This makes us closer to the reqwest::blocking API (and thus makes us closer
to being able to switch), avoids a lot of the callback madness and makes the
code easier to follow (except, unfortunately, all the curl wrapping).

The code doesn't handle all errors very nicely, but the old code didn't
either. On that matter, the new code is slightly better.