Build git-core in $OUT_DIR
Through a little bit of hacking around git-core's Makefile, we manage to
build in a different directory, which finally allows e.g. `cargo build` and
`cargo build --release` to build with the right flags without cleaning up
the source directory. `cargo clean` cleans everything up, too.

This also makes the submodule state non-dirty. \o/

This is still not 100% ideal, but it's already much better that what the
status quo was.