Avoid blocking when python code dies unexpectedly
Because we're sending an archive containing code via the python process's
stdin, we're currently relying on the rust part relaying its stdin to the
python process's stdin. But when the python process fails in an unexpected
way, the rust part stays stuck relaying its stdin.

Instead, we pass the rust stdin's file descriptor (a duplicate of it) to
the python process via an environment variable, so the rust process doesn't
need to do any relaying. This solves the problem as well as simplify the
code archive transfer.