Fail on `git fetch hg::url <sha1>`
`git fetch` normally only takes valid remote refs (listed by e.g.
ls-remote). But it also allows full sha1s, which it assumes are git
sha1s, so at the end of the transfer, it looks for that commit, and
subsequently fails.

So instead, we fail early (as nicely as we can), with an error message
similar to what git itself would print for other invalid refs.