Remove tag cache
It was added because things were slow to do through the fast-import
protocol, which is less true now, although the back-and-forth with the
helper protocol can be sub-optimal. It was also necessary to have a
cache per head because we used to have tags per repository, but nowadays
we only have an aggregate, so we don't need as much caching.

We could go ahead and just cache the overall tags and aggregate when new
changeset heads come in, but the performance is actually not bad without
the cache at all, and eventually, things will move entirely to rust,
making it faster. If a cache is needed, it's better to write it in rust
directly when that happens, than try to retrofit something that, all in
all, doesn't seem necessary anymore.