Die through Rust
It used to be sent to git's machinery because of the fast-import die
routine, which recorded a crash report with last commands run, etc.
but we stopped using that in 8049b5c4.

We can now do the opposite, which is to send git's crashes through Rust
panic code, which will get us "free" backtraces on crash.

We take on the occastion to normalize Rust panics to look more like git
failures, with an extra message to make them display the same message
as the python side does to print out tracebacks, and trigger a backtrace
when tracebacks are enabled.