Avoid issues when consolidating notes trees multiple times.
When fsck calls `dangling` for a notes tree, for_each_note does a
consolidation. When GitHgStore.close later writes out note trees if fsck
made adjustments, it consolidates again, but there are no additions after
the first consolidation, and the second consolidation reloads the original
tree, losing the modifications from before the call to `dangling` (and the
dirty flag).