Move bundle creation from python to Rust
The changegroup creation stays in python for now. For convenience, we allow
BundleWriter to handle raw changegroups.

We use the wire helper for the bundle creation because the import helper is
used by the python side to create the bundle, and the commands can't be
intertwined.

The protocol is honestly horrible, but it's meant as an intermediate step
only.