Make the wire thread optional
When starting the python command without a connection, we don't need the
wire thread running at all anymore, now that create-bundle is off of it.

That makes connect_with_main always take a connection, which is unsuitable
for git-cinnabar-wire, but that command has been useless ever since the
`connect` command was removed.