Rearrange how the bundle helper is spawned
The goal here is to spawn it only for things the Rust side can't handle
on its own. This means python now isn't involved when e.g. creating bundles
from a fresh clone.

This has the side effect of not creating .coverage profiles on some
CI jobs that now don't spawn python, and we need to mitigate that.