Remove the tools profile
After opt-level was accidentally dropped down to 0, it turns out normal
dev builds aren't that slow anymore at that level, so we can just use
the default now. The tools profile had been added to force an opt-level
of 0 on clippy builds because the optimization level made the jobs take
much longer.