Remove the preemptive buffering of ssh fetches
Back in the python code, this was mostly done to be consistent with the
HTTP handling, where it was done to work around some issues with proxies
cutting connections with no throughput for some time, which can happen
in some corner cases. Such networking connection problems are much less
likely to happen over ssh, and the buffering code is causing problems by
trying to read more from the pipe than what's available, and getting
dead-locked because of it. So just skip all that.