Change how git system config is initialized
Back when the config really only mattered for the wire connection part,
init_git_config was called without setup_git_directory ever being
called, and the config was preserved. But now that setup_git_directory
also happens, and because it resets the config, we lose the system
config if it's not in the same location as what git-core uses when
compiled.

We change the initialization to use an environment variable that
git-core will use during setup_git_directory, which also simplifies the
code on our end.