Implement a more generic way to recurse trees
This shares the recursion code between a normal tree iterator and a
diffing tree iterator. This also adds recursion support to
`merge_join_by_path`, which, while not used currently, will be useful in
the future.

I'm committing this with an exception for missing_docs because this is both
way too much and not enough at the same time, and I'm expecting massive
changes to the API, and probably the removal of most if not all these
traits that are added here.