Avoid infinite growth of the manifest cache during create_git_tree
When starting off from changegroupv1, the manifest cache is not used at all
during the manifest import phase. With changegroupv2, this might happen with
a very linear history, too. In that case, in create_git_tree, we end up
filling the tree because we hit the case where we never had a hit and always
get misses (create_git_tree avoids requesting the same tree twice), and that
can use a ton of memory uselessly, storing hundreds of thousands of items.
Hopefully, this doesn't impact performance in the changegroupv2 case before
of the slow cache warmup this causes.

The ideal thing would be for the cache not to be used at all, but this is good
enough for now.