In Git trees, parse file mode at the same time as looking for the space delimiter that follows it
It turns out to be adding a significant overhead. First, the conversion
in FromBytes, is slow. This can be alleviated with str::from_utf8_unchecked,
which makes the conversion as fast as a manual loop, but looking for the
length of the string to convert first is also a significant overhead (as
much as, or maybe even slightly more than str::from_utf8 vs.
str::from_utf8_unchecked).

Overall, this saves over 10% (!) of the time creating
changesets/manifests/files when bundling.